Replace lazy.rb with simpler deferred object
This stops the warnings from removing the object_id method.

Note that DeferredObject is not thread-safe, but this should not be a
problem as it is only being used to retrieve immutable data: the worst
case is that it is computed multiple times before being cached in the
instance variable.